backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
sse-starlette
google-generativeai
python-multipart
pydantic
Pillow
